Doctor

+++ ViewModels/PetDataViewModel.cs
@@ -8,7 +8,7 @@ namespace Veterinary.ViewModels
     public class PetDataViewModel
     {
         public string PetName { get; set; }
-        public byte PetAge { get; set; }
+        public int PetAge { get; set; }
 
         public Guid OwnerId { get; set; }
         public string OwnerName { get; set; } = string.Empty;
